Text parser
Added `core.text.Translator` class which gets string file in script
directory and loads them. Strings are saved into INI files. Each command
has to have a `help` key (which will be used for the help command).
If the string is gender-dependent, a `.f` or `.m` MUST be appended to
the key, `welcome_message.f` for example.
If the string contains variables, they should be surrounded in two
brackets: `((time)) since boot`.