Resolve some mypy errors
There is a bigger problem as we are using non-zero amount of dynamically
typed code. I think we may profit from type checks IN SOME PARTS OF THE
CODEBASE, but I don't believe we will ever be able to resolve it all,
without sacrificing readibility or without writing more code just to
make mypy happy.