Add own gettext implementation
Current translation system is not enough and has flaws, mainly poor
usability and mantainability. This is first of series of commits that
resolve this issue.

By extracting the strings into separate files we can keep the english
as default. As a result we can write the code in English and the
translations can be done independently.