gettext.py: Visit iterables
Resolves #96

New '_iterate()' function was added to simplify scanning of iterators.
It calls itself recursively over various forms of data structures,
making it more flexible.

Now also tuples, lists and dictionaries are visited.

This is still not perfect and may fail from time to time. It won't be
difficult to add more cases as they come.