Update gettext.py (#99)
* gettext.py: Visit iterables

Resolves #96

New '_iterate()' function was added to simplify scanning of iterators.
It calls itself recursively over various forms of data structures,
making it more flexible.

Now also tuples, lists and dictionaries are visited.

This is still not perfect and may fail from time to time. It won't be
difficult to add more cases as they come.

* gettext.py: Stop writing the string locations

While it may be helpful to know the translation source, it generates
a lot of noise.

Our codebase is not big and the wanted string can be found using both
standard UNIX command-line tools and integrated functions of editors and
IDEs.

* gettext.py: Add docstrings