Admin, Base: Use global LANGUAGES variable
Until now the supported languages were hardcoded on serveral places. Now
the bot will use the LANGUAGES variable specified in the core package.

Because the default is English, 'en' is not specified in that tuple.
Because of that, the functions that allow users to pick an interaction
language have to add the 'en' option where applicable.