Errors: Fix on_error handler
The documentation states that
> on_error will only be dispatched to Client.event().
>
> It will not be received by Client.wait_for(), or, if used,
> Bots listeners such as listen() or listener().

Because of that the 'on_error' listener has been moved up from errors.py
to the pumpkin.py.

The exception traceback is still printed to the stdout and is logged via
our logging mechanisms, even if it is not clear from this patch.
I'm not sure why it is happening, but it is what we want.