Add ISO 8601 as primary datetime parser
While fuzzy-matching dates is nice, it has a problem of being unable to
understand machine-friendly strings like the ISO 8601.

YYYY-xx-yy could be both YYYY-mm-dd and YYYY-dd-mm, depending on if the
day is <= 12 or not.

I also updated the tests so they are parametrized, and each of the
checks is done as separate test.