pie: Enable colored tables
Since the new Android app has been released a few weeks ago, we can
enable colored tables in the output. Support for it has been present for
several months now.

The colors are not visible on Android yet, but control characters are
not being displayed anymore. Most of the administration is usually done
on desktop.

We may add support for custom colors in the future, configurable by the
server administrators, to allow synchronization with role or server
colors.