pie: Force flush to stdout
We are printing log messages to stdout, along with saving them to a log
file. That is not a problem when the code is run directly in a terminal,
because they tend to display text immediately. However, when the stdout
is captured by systemd service, it can be held for quite a long time
(over ten minutes). By force-flushing the `print` statement we should
get nearly instant output.